PluginUpdate: send current resource state
This sends the current resource state in PluginUpdates. This was an idea
floating around and will get useful in the umh plugin (which will use
the current state to set environment variables).

While increasing the memory state, this makes it a lot easier for
plugins to consume updates (if they (also) need the current state).

While at it, wrap PluginUpdates into Arc.