build: add release target
This adds a release target (which for example is the source for the PPA
and the web page).

While this introduces a bit of copy/paste, IMO it is worth it. We have
projects that move that to a file list generation target calling make in
the Makefile, evaluating some PRESERVE flags, seding out directories in
the file list generation step... This is simple enough, KISS.

While at it use tar --transform, which is shorter and nicer. Also don't
include git/CI "meta data".