promoter: adjust resources
If the promoter is used for HA, especially for LINSTOR controller HA we
have some chicken/egg problem that LINSTOR itself can not bring up the
DRBD resource.

One could use drbd.service for that, but that just introduces another
dependency. drbdd/promoter can adjust such resources by itself.

Eventually this could be moved to a global level (i.e., having a
toplevel config entry for resources to start) as most plugins only make
sense if resources are started. On the other hand I assume that in
almost all cases LINSTOR will start them and therefore the promoter is
the special case.