plugins: split init/run
This splits the init phase (::new()) from the actual run phase.
This allows to do separate "initialization tasks" in new(). For example
the promoter plugin can use that to adjust the resources it monitors.

After that init phase we can then signal systemd that we are "READY=1"
and switch the service type to "notify" and have proper start
dependencies between services. This is the preparation step for that,
signaling is then almost trivial.

While restructuring the plugins also define a Plugin trait. This on one
hand allows us to define what we expect from plugins, and also allows us
to group them in one Vec for example.