events2: downgrade parse problems to debug
This is most likely a better idea anyways, and especially it generates
too much noise before we implement the umh plugin.

So all warnings about call/response (and path) are very well known and
harmless. Developers will run log level debug anyways.

We might change that back at a later stage. The idea was to make it
trivial for us to find all these version mismatch bugs between drbdd
and drbd-utils/events2. And in the end, after we got the basic set of
plugins and cover all events, then it really might be something we want
to warn about again.