introduce *PluginUpdate structs
In retrospect having these inline was a mistake to begin with. It was
mainly an easy way to avoid name clashes. These should be avoided by
moving the structs out of "drbd" to "events" and/or "plugins" eventually
anyways.

Having the declarations inline complicated parts of the code and this is
mainly a preparations step for the uml plugin. It is easier to use such
proper structs within macros that generate filter code.