make PluginUpdate structs matchable
This allows generation of sibling structs that are like the original,
but where all fields are optional. Additionally it generates a matches()
method, implementing the PartialMatchable trait, for the original
structs.

With that it is then trivial to create patterns that can be matched
against the original structs.

This could and maybe will be used in the promoter plugin, but got
developed for the umh plugin.

Co-developed-by: Roland Kammerer <roland.kammerer@linbit.com>