umh: implement user defined helpers
This plugin has 2 purposes:
- replace the helpers triggered by the DRBD kernel module by something
  more appropriate (e.g., that can run in a container).
- give users the flexibility to write their own rules.

This implements the second part, but there is no reason why we should
not give that to users already. Adding the first part can happen at any
times and will be pretty easy for drbdd, but will require changes to the
kernel module as well as to drbd-utils.