PluginUpdate: mv event_type to structs
This was structured the way it was because of how EventUpdate was
structured. For umh we will generate filter structs which we can then
use for matching. Moving the event type to the structs gives us
filtering event types for free.

This is a preparation step for umh, but merge it already so that others
notice the changed interface.

No functional change.