events2: add "path" information
This is mainly a preparation step for the prometheus plugin, which will
benefit from the IP information we then can expose.

While it looks weird that the get_path_updated() always returns None,
this is fine. We will internally split plugins into "change_plugins"
(current ones) and "events_plugins" (prometheus) soon. For now we only
have change_plugins and there we don't want to trigger updates for path
changes.

While at it refactor the Resource methods. This is new cleaner/easier
reusable by not handing in a type for some getters, but the actual
information (like a volume ID, or a peer ID). This still still is ugly
and will need some more love (e.g., replacing lists with maps,...).