plugins: split into "change" and "event" plugins
Currently we only send updates when important properties of a resource
change. This works very well for the plugins we currently have that do
something if a key property changes.

On the other hand we will have something like prometheus, that is
interested in certain details like the current number of out-of-sync
blocks. drbdsetup events2 in this case (a resync) even generates (rate
limited) events when these statistics change. But we certainly don't
want to send that to all plugins that are only interested in "major
changes".

So split the plugins and have:
- change plugins that get information whenever one of the key properties
  changed with that update (and the resource if they need it).
- event plugins that get the current Resource and type of event. As the
  resource is always a complete copy, the only type that should be
  checked is Destroy, as this might be used to for example delete an
  entry from an internal plugin cache.