prometheus: implement exporter
This is the first usable version of this plugin.

It misses peer path information (i.e., IP information). This is in
general useful, but can be added later.

Currently statistics are only updated during resync or if some other
DRBD event happens. This needs further discussion on how to handle that
on the DRBD side (poll like what we have in event2 --now --poll? or let
the kernel handle it and send some periodic statistics update?). Never
the less the plugin is already useful, it is a start.