promoter: wait till backing device shows up
It is possible that a promoter starts so early in the boot process that
the device/device link does not exist yet. LVM activation, udev,... For
pet resources like LINSTOR controller HA the admin could add a dependency
on a .device rule for example. More automated tools could add "wait for
this and that device" services to the "start = []" list. All in all
there is no harm to wait till the backing device shows up, especially
given that systemd is not strictly required.