promoter: make runner explicit
This is a preparation step the make the promoter plugin ways more
powerful by using systemd features. Splitting systemd/shell is easier
like this and I don't care for mixed shell/systemd anyways.

The only reason for me to keep "shell" is that the promoter might be
used on Windows one day. The promoter plugin is interesting because
there obviously is no Pacemaker for Windows. In this case
"plugin::system" would learn to use "the windows shell". Or we would add
a specific enum variant for whatever Windows uses as service handler.