promoter: wait a bit after stop
If we failed to start and revert or we lost quorum it is a good idea to
give some other node the chance to start the services.

In case we know that we stop intentionally (plugin stop/drbd-reactor
service stop) we can skip that. Even with a small number of promoters
serialized stop would sum up quickly. Usually plugins just stop, the
only one that might do "cleanup" is the promoter, so no need to touch
the core for that.