promoter: compat with older systemd, make promotion a dependency
Older systemd does not understand ExecCondition and ignores it. If
auto-promote is disabled, which it should, then the promotion silently
fails (actually is not executed).

I want that to be a StartExec anyways, which is also understood by older
systemd. In general and now on newer systemd, if ExecCondition fails,
also ExecStart is ignored. So while unlikely that the same command
fails and then succeeds in the override, and even though "drbdsetup
primary" is idempotent, just unset the original ExecCondition.

I think this should be a ExecStart in a promoter use case anyways, but
compat to older systemd matters as for example Proxmox stable currently
has a too old sytemd (Debian 10).