ctl: use tempfile not in snippets
So far a directory in the snippets directory was used as scratch space.
If we use inotify based reloading and monitor the directory, this
results in daemon reloads. Use a regular tempfile. This is named and
probably could be guessed. But all in all the same guess/override
attacks existed before as well.