promoter: fix ocf regex
So far our regex did not correctly match OCF agents with a '-' in their
name, which actually is the preferred naming scheme nowadays. This now
uses non-space characters, we assume no funny vendors/agents with a ':'.

It also introduces a space between vendor and "the rest" (i.e., starting
with the instance name).