ensure reload updates don't trigger promoter early
When reloading drbd-reactor, we resend the current observed state
to new plugins, so that these plugins know about existing resources.

When doing this the naive way, we may trigger the promoter plugin too
early: The first thing we send is the Resource update, which already
contains the current may_promote state. This can trigger a promotion
on a resource that does not contain any devices. This means that the
delay logic for diskfull vs diskless resources would not have the
necessary information and use no delay at all.

This commit ensures that for every existing resource:
* We announce the resource exists with (may_promote: no, role: Secondary)
  This is the default for newly created resources.
* We announce devices, connections, etc..
* We create a change event on the resource, setting the actual state
  of the resources as memorized by the reactor core.