promoter: persist on-failure-action
This tries to persist the journal after writing that an
on-failure-action is going to be executed. This then helps to find out
why a node was shut down (e.g, echo b /proc/sysrq-trigger/).

While at it also remove setting "true" as default action. That was kind
of a hack anyways, let's remove this pointless hack.