containers: use cargo from ubi8
The ubi8 images have a working cargo/rust toolchain, no need for rustup.

For the internal build system we override the BUILDER which has a
rustup toolchain, so make sure we honor $HOME/.cargo/env

All shell variants are for handling ~/.cargo/env are pretty ugly, so
just keep it as is but make sure it exists.

Funny enough the '.' for sourcing the file behaves differently in
different shells for non existing files. busybox and dash error out;
bash and zsh behave a bit more like one would expect and continue to
execute. Reproducer: '. /doesnotexist ; echo here'