promoter: deprecate on-stop-failure; introduce on-drbd-demote-failure
The on-stop-failure action actually never worked. Without going into
details a stop failure is pretty much impossible to detect with systemd.
So let's take one step back, what do we actually want to achieve? We
want to reboot/force reboot/poweroff/... when we are in a stuck
situation.

We expect that services propagate their failure to the target, which
then stops and starts services. We are stuck if we can not demote the
involved DRBD device. Because something is holding it open, drbdsetup
hangs, whatever. And that is something we control.

The user can now set on-drbd-demote-failure to a string as defined for
systemd "FailureAction=" (systemd.unit(5)).

As we have systems out there that already have set on-stop-failure, the
best we can do is to ignore that setting.