core: dump config on debug level
This makes it a bit easier to see what configuration is currently used.

It is not perfect as there is also the reload logic which might keep
started plugins as they are (if their ID did not change; no matter how
the rest of their config changed), which is something you don't see in
this config dump. Anyways, that is the minimal version I accept after
the RC, the rest (e.g., change the Plugin trait) is something for the
next release.