promoter: fix stop-services-on-exit
Commit 27d798e0 introduced setting Before=drbd-reactor.service in the
generated drbd-services@ target overrides. While a good idea in general
(see that particular commit), it also can create a situation where the
user issues a `systemctl stop drbd-reactor` (expecting switch-over),
which then blocks as:
drbd-reactor.service want to shut down and promoter wants to `systemctl
stop` the target unit. But this contains the mentioned Before= and wants
drbd-reactor to be gone first. Blocks and timeouts.

We can borrow Pacemaker behavior by writing the Before= in a extra
override and when the plugin detects it is responsible for shutdown, it
then can:
- rm that particular Before= override file.
- systemctl daemon-reload
- systemctl stop the target (that now does not have the Before=
  dependency on drbd-reactor.service)

All fine, but with that we saw another behavior becoming a show stopper:
systemd.mount(5) automatically adds a Before= dependency on
local-fs.target unless "nofail" is specified. That in combination with
what was described above made systemd fail with a cyclic dependency on
"reboot". That can be solved by setting "nofail" as an option, or what
looks even a slightly better in our case by setting
"DefaultDependencies=no".

This also introduces by number sorted reactor overrides, but keeps the
general reactor.conf (we could start with 50 as default). This makes
upgrades easier as we don't need to fear a leftover reactor.conf that
overrides all the others. Ordering should not be necessary, as I don't
expect one override to override settings in another one. If that ever
happens we need to do that in the numbered ones (this already prepares
for that) accordingly and be aware that rector.conf overrides rules them
all.