promoter: systemd escape template names
So far resources that needed escaping did not work (e.g., 'rck-1'). This
escapes such names where needed. This includes the target unit name
(i.e., the "handle" to start and stop), as well as inter dependencies,
and OCF template names.

It also adds that escaping to drbd-reactorctl which directly calls
systemd-escape.