core: preserve pds/paths on conn updates
So far connection updates as triggered from the core event handling
overwrote the existing connection with the current event (e.g., Change).
The input connection struct never contains paths/peerdevices as these
are events on their own, but are stored in the connection struct. So we
overwrote the pds/paths in the existing connection with empty pds/paths.

This could be fixed in update_connection, but in theory this function
could receive a connection with existing pds/paths. So it is better to
assume that structs are complete at this point. So fixing this at a
"higher level" in get_connection_updated seems more appropriate.