ctl: add evict command
This command can be used to evict ("move away") promoter plugin managed
resources.

It checks if it is a promoter resource, and if it is DRBD Primary on the
local host. If that is the case the resource is disabled (and the
systemd target gets stopped). The we check if another node took over and
print status information accordingly. In any case the plugin is
re-enabled so that the node can take over the resource again.

Caveats: While we can iterate over individual plugins, we can only
disable all plugins within a config file. So if one really mixes a
promoter and let's say a prometheus plugin in one config snippet, both
the disabled (and re-enabled). Second corner case is a promoter that is
responsible for multiple resources. Bonus points for multiple promoter
being responsible for multiple resources in one snippet. This should be
very rare and we have a check in place (maybe we should make it the
global default?). The user has to override with '--force' in that case
and then one gets what one asked for (multiple disable/enables).