ctl,evict: order files by demote failure action
This is just some convenience for the user. I assumed a mixed setup of
resources that have a drbd-demote-failure action set and ones that
don't. If then one with an action fails, it might reboot the cluster. It
would be nice if we evict the ones without such an action first.