ctl,evict: use systemd, allow maintenance
The current version has one problem: If a on-drbd-demote-failure action
is triggered between the time the plugin (config) is disabled (i.e.,
actually moved on the file system) and then it would get re-enabled,
then the config stays disabled. This is avoided by using "systemctl mask
--runtime" (thanks Moritz for the idea). If now things go wrong and the
node reboots, the mask (in /run/systemd) get cleared and after a reboot
things are back to normal.

While at it also implement a stricter version of evict. As of now the
resources got evicted, but there were no means in place to avoid that
resources fall back. This is a valid behavior, like when you just want
the resource to be active somewhere else. But for maintenance one wants
to avoid any fallbacks. This can be done by "evict --keep-masked". Then
the target unit stays masked and the resource will not be active on that
node. This uses "systemctl mask --runtime" so that is cleared after a
reboot (e.g., kernel upgrade + reboot). It can be manually cleared via
"evict --unmask".