build: allow signal-hook >= 0.1
This is required for "proper" Ubuntu Jammy PPA building, which has
version 0.3 of that library. Unfortunately we want to support both Focal
and Jammy, and Focal has 0.1. We can use the signal definitions from
libc (they got moved to consts), and change the mutability. This now
generates a warning on Focal, but that is okay.

As this is in a part we don't touch too often I hope we will not break
it in the future.

While at it, run a "cargo update".