promoter: implement preferred nodes
This allows setting a list of preferred node where a resource should be
active.

In general I'm not a huge fan of that feature, as every node has to be
able to run all resources anyways. You are HA or you are not.

This works by calculating a sleep delay that is added depending on the
node's position in the preferred nodes list.

If a peer joins the DRBD resource (e.g., comes back after a failure) and
that peer's disk is UpToDate and a preferred node compared to the one
that has the resource active currently, then the local node stops the
services. The preferred node then sleeps shorter and will take over the
services.