promoter: rm path dep for promote service
As is this makes ways more troubles than it solves. adjust and template
generation could fail because we can not access the res file and fail to
execute "drbdadm sh-ll-dev".

We could have relaxed that and only write the devices if we can find
them. Still, this would break cases where storage is moved around (or
linstor toggle-disk).

In general:
- the res file exists, and reactor will do the initial adjust and wait
  for backing devices.
- if the res file does not exist (e.g., LINSTOR satellite and its "rm
  res files on startup" logic), then we try, fail, and simply assume it
  is the job of that component (e.g., LINSTOR) to bring up the
  resources.