events2: make parser more robust
Recently we had the problem that a newer drbd-utils version introduced a
new keyword. The result was that the parser stopped and returned an
error.

This error was ignored, but still, also the complete line was ignored.
And that line also contained the "may_promote" keyword. The result was
the resources never promoted.

Now we just write a warning to the log, but parse all the keywords we
can.