promoter: implement on-quorum-loss policy
This is the short version of what can be found in promoter.md, for more
details see that document.

There are services that take a long time to start and immediately
shutting down the service on one node might not preferable. It might be
the case that for example the other nodes also did not create a new
partition, but when they join again, we have quorum again and then we
can just "thaw" the "frozen" service.

If we are the Primary and lose quorum we freeze. If we gain quorum
again, we thaw. If the other partition selected a new Primary and they
connect, we get forced to Secondary and stop the service. Last, if the
replication net was lost and we are now the frozen Primary, and the net
comes back, we want to make sure that "randomly" connected Secondaries
wait a bit till they start the service as they might connect to old
Primary shortly and then we would just need to thaw the service instead
of starting it on the Secondaries selecting a new Primary immediately.