promoter: warn if mount unit is topmost unit
This has been discussed in blog posts in detail, the short version is:
When systemd stops the mount unit there might still be users outside the
control of systemd (some user doing things with the FS). This is bad,
because the umount will fail. In such a case use a OCF FS RA as it does
some extra work to kick users (e.g., fuser -k).

Mount units are fine, but only if there is a service on top of them that
is also controlled by reactor/systemd, then systemd will make sure the
users of the FS go away by itself.