ctl: fix status without res filter
Unfortunately as is this gives us:
Traceback (most recent call last):
  File "/usr/sbin/drbd-reactorctl", line 778, in <module>
    args.func(args)
  File "/usr/sbin/drbd-reactorctl", line 374, in status
    if len(resources) > 0:
TypeError: object of type 'NoneType' has no len()

I assume that is because getattr in this case "does not work", as "resource"
exists, but is a NoneType.  Fix that by setting a default. Then all variants
("... status", "... status foo", and plain "drbd-reactorctl") work as expected.