promoter: call systemctl freeze/thaw for every unit
This makes it a bit easier to filter units we don't want and we can
print better warnings. So far if 1 unit "failed" to freeze/thaw because
there is no active process in it, we basically overreacted by telling
the user freeze/thaw failed, even if that was expected. Give the user a
hint.