ctl: rewrite in Rust
The trigger was that the toml library we used in the Python code has an
ugly bug with multi-line literal strings and we would like to use these
to avoid overlong lines in our documentation. The bug in that lib is
known and fixed, but exists in almost all versions distros ship.

Overall advantages:
- toml lib that is correct
- this is already a Rust project
- reduced dependencies (Python and extra py libs not shipped by some
  distros, extra program for tab completion generation)