ci: test compatibility with different rust version
When building packages, we try to use the native toolchain for the
distribution, if it is available. Sadly this means we need to support
a range of rustc versions, where some features or library functions
are not yet available.

To prevent any surprises when cutting new releases, test against
all rustc versions we want to support in CI. We use cargo check
to speed up this step as much as possible. If that passes, we
can be pretty sure it will also compile and work for real.