daemon: ensure logging is available early
We had an issue with crashing drbd-reactor processes that do not print any
failure information. This may have been caused by a late initialization of
the logging infrastructure.

With this commit we first read the config, making sure any errors
are appropriately printed on exit. Immediatly afterwards, we set up the global
log sink.

This commit also ensures that any errors during signal handler registration
are picked up, as well as given some additional context when checking the
DRBD version.