core: drbd checks only if we have promoters
So far we ran the drbd version checks once at startup no matter which
plugins are active. The disadvantage is that we actually only need to be
careful about DRBD versions for the promoter plugin. That might change
in the future (e.g., umh when it does the kernel side handlers) but as
of now this really is only interesting for promoters.

In a k8s case we run promoter but only to export DRBD metrics where the
exact DRBD version does not really matter that much. In this case it is
fine to run on "too old" DRBD.

Check if there are promoters involved and fail hard if the config is
"wrong" on reactor startup. Ignore new configs if there are promoters
that then don't fulfill the minimal version.