clt,start-until: simplify instructions
If the user is done with whatever she wanted to accomplish with
start-until there is no need actually stop the target and then start the
whole target by enabling the target from scratch. For example there
might be a database that takes ages to start and is already started.
Then it is better to just start the rest of the target.