prometheus: expose drbd_version
This was a request and is pretty self explanatory. Currently this has
a limitation: the version is retrieved once, and then not updated till
the plugin is restarted. DRBD kernel module reloading will require work
on events2 anyways. Even if we have that in place we would need to
re-read the version in the prometheus plugin. When we have the events2
changes in place, we could think about a special PluginUpdate variant
that we send to plugins to "reevalutate the world", whatever that is in
the plugin's context.