promoter,ocf: fix env for old systemd
They way we generated the Environment variables was fine for modern
systemd, but did not work for CentOS7-alike distributions. We can fix
that by doing some 'systemd-escape'. We need to be careful to use a more
relaxed version that for example does not replace '/'. This could be
optimized, but resembles what systemd-escape does closely and should be
easy enough to read.