prometheus: start server in plugin constructor
The main point where the prometheus plugin could fail was during the initial
bind() of the socket. The core loop does not wait for any plugin threads before
shutting down in case of errors, which means we might miss a crucial log
message.

By having the bind() be part of the plugin setup, we are guaranteed at least a
log message.

While at it, also clean up the code somewhat, making sure the socket it shut
down when dropping the plugin struct.