ctl: ignore already unmasked targets
During eviction, if a resource is already primary on another node before
the eviction starts, the service is never masked. During clean-up, we then try
to unmask the service. Up to now this step failed because the mask file did not
exist.

This is fixed by ignoring non-existent mask overrides when unmasking services.