core,plugins: auto generate IDs
So far every plugin instance could set an optional ID. The ID was then
used as "handle" to decide which plugins need to stop, start, stay alive
when drbd-reactor gets reloaded. This worked, but is not what users
expect and we can do better. For example if a user edited the toml
snippet for a plugin and did a reload, the old instance was kept alive
as the ID did not change.

By making the configuration structs trait Hash we have a very natural
ID, namely the hash of the config and we can use that instead of manual
IDs. The advantage is obvious, if the content, even better, the semantic
of a configuration snippet changes, it has a new hash and the old
instance gets stopped and the new started on drbd-reactor reload.