core,plugins,ctl: deprecate IDs
As we now have auto generated IDs (i.e., hashes of the configuration) we
can deprecate the "id" field in the configuration structs. As we want to
write some warning message to the log if "id" is set, we can not set it
"#[deprecated]".

This also touches some bits in drbd-reactorctl to keep it useful without
the old manual IDs. While at it, it also improves "ls" a bit by
providing some additional information if possible like the address of a
Prometheus plugin, or the start list of a Promoter.