core,daemon: don't start if tty
We had a case where a user wanted to start "drbd-reactorctl", but hit
<enter> after the initial completion too early. This might have
"unfortunate" consequences. On the other hand we want to have the option
to still start the daemon intentionally for some debugging reasons. This
is a pretty internal use case, I did not even care to document that
option, the hints printed should be enough.