agentx: implement plugin
Most of the work happened in https://crates.io/crates/agentx which also
gives some pointers to to the protocol involved. In short we can use
net-snmp, the de facto standard, and configure it to accept connections
from AgentX conforming sub agents like this one.

The implementation tries to be rather robust, for example it should
survive disconnections from the snmpd well and will reconnect.