prometheus: ensure default works with any IP family
If a user tries really hard by setting ipv6.disabled=1 on the kernel
command line, our existing default "[::]:9942" address no longer works.

Instead, we create a fallback that uses "[::]:9942" and falls back to
"0.0.0.0:9942". It depends on the feature of TcpListener to try bind
socket addresses in the provided order.

If a user does specify an address nothing changes: if they can set the
address, it's reasonable that they know which format works for them.