events2: wait for drbdsetup command on exit
We already implemented a kind of "clean up" by killing the child process. But
we never actually waited on it to exit. This leaves behind a zombie process
every time drbdsetup exits unexpectedly.

So we just add a wait() call on our drop guard: the process should have been
killed in any case, so we should not be waiting long.