events2: exit reader thread on error
We saw that when drbdsetup is respawned we get a pipe between
drbd-reactor and drbdsetup that is still inoperable. In such a case we
see "Broken pipe (os error32)" in the logs.

Exiting the thread here should be fine. I assume not existing is a
leftover from when events2 handling was in an early state where
respawning was not possible, so we had to keep things running at all
costs.

The idea is that exiting the thread is good enough and that we exit the
whole process_events2 function because we also fail in the read loop
below.