promoter: systemctl daemon-reload only once
So far we assumed that ::new() is cheap and we don't really need to care
about it, but with hundreds of promoter plugins things accumulate. We
can easily postpone all the daemon-reloads. We also can improve on
telling systemd our status (i.e., sd_notify()) which will happen in an
extra commit.