promoter: use Stdio::null() for systemctl commands
On startup, but in theory in general, we might run lots of systemctl
commands in parallel (systemctl start drbd-services@.target in our
case). Unfortunately systemd wants to set watches for "password files".
This then fails and in the worst case systemd DoSes itself. We can avoid
that by using null as stdin.

Co-developed-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>