promoter: do not generate ExecStart override
Historically we always overwrote ExecCondition with ExecStart as some
older (e.g., rhel8/debian bookworm) systemd version don't support it.
Actually ExecCondition is exactly what we want here so it would be nice
to not overwrite it on systems that can handle it. drbd-utils now has
functionality that generates the drbd-promote@.service template in a way
that fits the base OS (i.e., systemd version). Old systems get a
ExecStart in the template, new ones a ExecCondition, no need to
overwrite it in reactor::promoter anymore.