promoter: start services on activation
We had a scenario where a user used "drbd-reactorctl disable --now" but
demotion failed for whatever reasons. Then they re-enabled it and were
surprised that services did not start. This was because the DRBD
resource was still Primary, so we never saw a fresh "may_promote". In
run() we now start the target if we are locally Primary, this might help
in the described scenario and should not cause any harm in general.