promoter: add preferred-nodes-policy
So far when preferred-nodes were set they also were taken into account
when a service was already running on a lower priority node and the
higher priority node joined later. In this case the lower priority node
stopped the local service and (most likely) the higher priority node
took over.

It is a valid use case not wanting to fall back if services are running
and a higher priority node joins. One might want to keep the running
services on the lower priority node. This is now implemented via
preferred-nodes-policy, which can be "always" or "start-only".