example: add tmpcontroller.sh
This is a script that allows to start a temporary controller from a
backing device. It executes various checks and then creates a RO overlay
for /var/lib/linstor/.

Consider it as a PoC, always adapt it to your use case.

Note that this is a last resort to get resources up, but never ever use
the temporary DB to modify actual state. For example if you create a new
resource in the temporary DB, then actual backing devices are created on
the host. If you then roll back to the actual DB (after fixing things),
then a) the modifications in the DB are lost and they b) might be in the
way when you create a new resource. This includes but is not limited to:
backing devices, DRBD resource name, minor numbers, ports,
snapshots,...

Use this with care and if you know what you are doing.