promoter: generate stronger promote service dep
We use the ExecCondition= in the drbd-promote@.service to "silently" try
promotion on systems where possible. This causes an unintended
interaction with the default Requires= dependency.

To quote from man systemd.unit:

> Note that this dependency type does not imply that the other unit always
> has to be in active state when this unit is running. Specifically:
> failing condition checks (such as ConditionPathExists=,
> ConditionPathIsSymbolicLink=, ... — see below) do not cause the start
> job of a unit with a Requires= dependency on it to fail. Also, some unit
> types may deactivate on their own (for example, a service process may
> decide to exit cleanly, or a device may be unplugged by the user), which
> is not propagated to units having a Requires= dependency. Use the
> BindsTo= dependency type together with After= to ensure that a unit may
> never be in active state without a specific other unit also in active
> state (see below).

To fix this we always generate a "strong" (i.e., BindsTo) dependency on
drbd-promote@.service.

This still uses vecs where we could use a string, but tinytemplate
handles empty vecs nicely, so we avoid touching the template too much.
And it keeps it a bit more general.